Reorganize Showcase's dependency registration
Simply refactor and re-organize the logic to register the Showcase's dependencies registration in the Microsoft DI container.
