+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Web/Program.cs
@@ -1,17 +1,7 @@
-using EntityFrameworkCore.UseRowNumberForPaging;
-using Hangfire;
-using Microsoft.EntityFrameworkCore;
 using NLog.Web;
-using Smart.Design.Razor.Extensions;
+using Smart.Extensions.DependencyInjection;
 using Smart.FA.Catalog.Shared.Security;
-using Smart.FA.Catalog.Showcase.Domain.Common.Options;
-using Smart.FA.Catalog.Showcase.Infrastructure.Data;
-using Smart.FA.Catalog.Showcase.Infrastructure.Mailing.Inquiry.SmartLearningTeam;
-using Smart.FA.Catalog.Showcase.Infrastructure.Mailing.Inquiry.Trainer;
 using Smart.FA.Catalog.Showcase.Web.Extensions;
-using Smart.FA.Catalog.Showcase.Web.Options;
-using Smart.FA.Catalog.Showcase.Web.Services.Trainer;
-using Smart.FA.Catalog.Showcase.Web.Services.Training;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -21,29 +11,10 @@ builder.Host.UseNLog();
 // Add services to the container.
 builder.Services.AddRazorPages();
 
-// Add localization.
-builder.Services
-    .AddFluentEmail(builder.Configuration)
-    .AddShowcaseLocalization()
-    .AddTransient<ITrainingService, TrainingService>()
-    .AddTransient<ITrainerService, TrainerService>()
-    .AddTransient<ISmartLearningInquiryEmailService, SmartLearningTeamInquiryEmailService>()
-    .AddTransient<ITrainerInquirySendEmailService, TrainerInquirySendEmailService>()
-    //.AddHangfire(builder.Configuration)
-    .AddMemoryCache()
-    .AddSmartDesign();
+// Add every dependencies required by the Showcase web app.
+builder.Services.AddDependencies(builder.Configuration);
 
-builder.Services.Configure<MinIOOptions>(builder.Configuration.GetSection(MinIOOptions.SectionName))
-    .Configure<FluentEmailOptions>(builder.Configuration.GetSection(FluentEmailOptions.SectionName))
-    .Configure<InquiryOptions>(builder.Configuration.GetSection(InquiryOptions.SectionName));
-
-builder.Services.AddDbContext<CatalogShowcaseContext>((_, efOptions) =>
-{
-    efOptions.UseSqlServer(builder.Configuration.GetConnectionString("Catalog"), options => options.UseRowNumberForPaging());
-});
-
-builder.Services.AddHttpContextAccessor();
-
+builder.Services.ConfigureShowcaseOptions(builder.Configuration);
 
 var app = builder.Build();
 
@@ -74,7 +45,5 @@ if (!app.Environment.IsProduction() && !app.Environment.IsEnvironment("PreProduc
 {
     await app.BootStrapAsync();
 }
-
-//app.UseHangfireDashboard();
 
 app.Run();
